Split generic behavior method metadata tests
